Add models (#21)
In progress - lots of work on data cleaning, exploration and visualization, and run through untuned classification models. There's a TODO list at the top of things we will be trying to improve the model.

Best performers, untuned, can be found at tm_untuned.pkl and rf_untuned.pkl. Inputs and types are listed at the top of the notebook.